数据库表修改后，相应修改程序； 热量表表号组成原则：14位数字 = 6位厂家代码 + 8位数字
git-svn-id: https://192.168.1.124:8443/svn/delusoft/trunk/qtverify@300 62fcab80-fd93-7748-8c49-e77283212eb3

+++ source/include/weightmethod.h
@@ -103,7 +103,7 @@ public:
 	bool m_autopick;          //是否自动采集
 	int m_flowPointNum;       //流量点的个数
 	int m_exaustSecond;		  //排气时间
-	int m_totalFlag;          //1:总量检定；2:分量检定
+	INT16 m_totalFlag;        //1:总量检定；2:分量检定
 	int m_standard;           //表规格
 	int m_model;              //表型号
 	int m_meterType;          //表类型
